Change FillCreateItemGroupEvent to CreateItemGroupArrangerEvent
